Feature/redis-queue

+++ src/main/java/io/hhplus/concert/domain/token/TokenInfo.java
@@ -5,6 +5,6 @@ import io.hhplus.concert.common.enums.TokenStatusType;
 public record TokenInfo(
         String token,
         TokenStatusType status,
-        Integer position
+        Long position
 ) {
 }
